fix(graphql): dissociate phase blocks fetch from phase

+++ app/nemeton/range.go
@@ -3,7 +3,7 @@ package nemeton
 // BlockRange Represents a blockchain block range.
 type BlockRange struct {
 	// The block height the range begin, inclusive.
-	From int `bson:"from"`
+	From int64 `bson:"from"`
 	// The block height the range end, inclusive.
-	To int `bson:"to"`
+	To int64 `bson:"to"`
 }
